requirements.txt
flask
flask-sqlalchemy
flask-login
psycopg2-binary
flask-sock
python-dotenv
vonage
websocket-client
gunicorn
werkzeug
gevent
